move data to embedding_data.py; fix inference
